+++ internal/impl/sentry/processor_capture.go
@@ -18,8 +18,8 @@ const (
 
 func newCaptureProcessorConfig() *service.ConfigSpec {
 	return service.NewConfigSpec().
-		Version("4.14.0").
-		Description("Captures log events from messages and submits them to [Sentry](https://sentry.io/).").
+		Version("4.16.0").
+		Summary("Captures log events from messages and submits them to [Sentry](https://sentry.io/).").
 		Fields(
 			service.NewStringField("dsn").
 				Default("").