+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+## 4.16.0 - 2023-05-28
+
 ### Added
 
 - Fields `auth.user_jwt` and `auth.user_nkey_seed` added to all NATS components.
@@ -18,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - The `aws_s3` input now adds `s3_version_id` metadata to versioned messages.
 - All compress/decompress components (codecs, bloblang methods, processors) now support `pgzip`.
 - Field `connection.max_retries` added to the `websocket` input.
+- New `sentry_capture` processor.
 
 ### Fixed